fix withdraw to float

+++ internal/models/Order.go
@@ -35,11 +35,11 @@ func GetResponseGetOwnerOrders(source []Order) (result []ResponseGetOwnerOrders,
 
 // ResponseGetBalance описывает элемент ответа пользователю на получение суммы его баллов и суммы списаний.
 type ResponseGetBalance struct {
-	Current   int `json:"current"`
-	Withdrawn int `json:"withdrawn"`
+	Current   float64 `json:"current"`
+	Withdrawn float64 `json:"withdrawn"`
 }
 
-func GetResponseBalance(current, withdrawn int) (resp ResponseGetBalance, err error) {
+func GetResponseBalance(current, withdrawn float64) (resp ResponseGetBalance, err error) {
 	resp = ResponseGetBalance{
 		Current:   current,
 		Withdrawn: withdrawn,
